Re-make the canonical loc the strictest
This essentially re-instates 9c1dfefb30a71ad3a48a3ecf0651428e8fbe282e,
reverting bd4b8b9d3933522ed70f77da297525fa1f6b7b9a which reverted that.

It seems that `Symbol::loc()` was updated since 9c1dfefb to treat the
_last_ loc in `locs_` as the canonical loc. In 9c1dfefb, it was the
_first_ loc that was treated as canonical.

Since the intent is that the new loc should be made the canonical loc
iff it is stricter than the old canonical loc, we have to switch around
adding to the front and adding to the back here.